requirements.txt
Django==2.1.5
Pillow==5.4.1
django-materializecss-form==1.1.11